+++ web/app/scripts/compiled-es5/directives/proportion-field.js
@@ -10,18 +10,23 @@ var _proportionField = require('./../services/proportion-field.js');
 
 var _proportionField2 = _interopRequireDefault(_proportionField);
 
+var _integerField = require('./integer-field.js');
+
+var _integerField2 = _interopRequireDefault(_integerField);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var moduleProperties = {
     moduleName: 'incuna-surveys.proportion-field-directive'
 };
 
-var _module = _libraries.angular.module(moduleProperties.moduleName, [_proportionField2.default.moduleName]);
+var _module = _libraries.angular.module(moduleProperties.moduleName, [_proportionField2.default.moduleName, _integerField2.default.moduleName]);
 
 _module.directive('proportionField', [_proportionField2.default.componentName, function (ProportionField) {
     return {
         restrict: 'A',
         scope: {
+            form: '=',
             model: '=',
             options: '=proportionField'
         },
@@ -46,6 +51,9 @@ _module.directive('proportionField', [_proportionField2.default.componentName, f
             });
 
             scope.$watch('model', function (values) {
+                if (values === null) {
+                    scope.model = {};
+                }
                 scope.total = ProportionField.calculateTotal(values);
                 if (values) {
                     ProportionField.addPercentages(scope.fields, values, scope.total);