+++ web/app/scripts/compiled-es5/services/proportion-field.js
@@ -30,11 +30,11 @@ _module.service(moduleProperties.componentName, [function () {
     };
 
     this.calculateTotal = function (values) {
+        if (!values) {
+            return 0;
+        }
         return Object.keys(values).reduce(function (total, key) {
-            if (!values[key]) {
-                return total;
-            }
-            var value = values[key] && parseInt(values[key], 10);
+            var value = parseInt(values[key], 10);
             if (Number.isNaN(value)) {
                 return total;
             }