+++ web/app/scripts/src/services/fieldsets-parser.es.js
@@ -53,14 +53,16 @@ module.service(moduleProperties.componentName, [
         };
 
         this.addFieldErrors = function (fields, errors) {
-            fields.forEach((fieldset) => {
-                const key = fieldset.templateOptions.id;
-                const fieldsetErrors = errors[key];
-                fieldset.fieldGroup.forEach((field) => {
-                    const fieldError = fieldsetErrors && fieldsetErrors[field.key];
-                    field.templateOptions.fieldOptions.errors = fieldError;
+            if (fields && errors) {
+                fields.forEach((fieldset) => {
+                    const key = fieldset.templateOptions.id;
+                    const fieldsetErrors = errors[key];
+                    fieldset.fieldGroup.forEach((field) => {
+                        const fieldError = fieldsetErrors && fieldsetErrors[field.key];
+                        field.templateOptions.fieldOptions.errors = fieldError;
+                    });
                 });
-            });
+            }
         };
     }
 ]);