+++ web/tests/dummy-spec.js
@@ -3,12 +3,10 @@ describe('test', function () {
     beforeEach(function () {
         fixture.setBase('tests/api-description');
     });
-    
 
     it('should load the api description', function () {
         const testFixture = fixture.load('dummy.json');
         expect(testFixture.OK).toBe('OK');
     });
-    
 
 });